requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.49
asyncpg==0.30.0
psycopg2-binary==2.9.12
alembic==1.18.4
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.6
boto3==1.34.53
Pillow==12.2.0
python-dotenv==1.0.1